Maign | serviceType and service delete handled for dependencies

+++ src/services/database/typeMaster/generalMaster/serviceType/delete.ts
@@ -1,15 +1,34 @@
-import prisma from "../../../database.js";
-import throwDatabaseError from "../../../utils/errorHandler.js";
+import { StatusCodes } from 'http-status-codes';
+import APIError from '../../../../errors/APIError.js';
+import prisma from '../../../database.js';
+import throwDatabaseError from '../../../utils/errorHandler.js';
 
 async function deleteServiceTypeDB(id: string) {
   try {
-    const deletedService = await prisma.serviceType.delete({
-      where: {
-        id: id,
-      },
-
-    });
-    return deletedService;
+    const deleteTransaction = await prisma.$transaction(
+      async (prismaTransaction) => {
+        const dependency = await prismaTransaction.serviceType.findFirst({
+          where: { id },
+          include: {
+            service: true,
+          },
+        });
+        if (dependency && dependency.service.length > 0) {
+          throw new APIError(
+            'service type has dependencies ',
+            StatusCodes.BAD_REQUEST
+          );
+        } else {
+          const deletedService = await prisma.serviceType.delete({
+            where: {
+              id: id,
+            },
+          });
+          return deletedService;
+        }
+      }
+    );
+    return deleteTransaction;
   } catch (err) {
     if (err instanceof Error) {
       throwDatabaseError(err);
@@ -17,15 +36,15 @@ async function deleteServiceTypeDB(id: string) {
   }
 }
 
-async function deleteServiceDB(prismaTransaction:any,id: string) {
+async function deleteServiceDB(prismaTransaction: any, id: string) {
   try {
     const deletedService = await prismaTransaction.service.delete({
       where: {
         id: id,
       },
     });
-    console.log("delete db layer")
-    console.log(deletedService)
+    console.log('delete db layer');
+    console.log(deletedService);
     return deletedService;
   } catch (err) {
     if (err instanceof Error) {
@@ -42,7 +61,10 @@ async function deleteUncheckedServices(
   try {
     for (let exisitingId of exisitingServicesId) {
       if (!updatedServicesId.includes(exisitingId)) {
-        const deletedService = await deleteServiceDB(prismaTransaction,exisitingId);
+        const deletedService = await deleteServiceDB(
+          prismaTransaction,
+          exisitingId
+        );
       }
     }
   } catch (err) {
